requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-admin-interface==0.19.0
django-colorfield==0.6.3
django-cors-headers==3.11.0
django-filter==21.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-mssql-backend==2.8.1
django-pyodbc-azure==2.1.0.0
django-restframework==0.0.1
django-taggit==2.1.0
django-timezones==0.2
django-utils==0.0.2
djangorestframework==3.13.1
frozenlist==1.3.0
future==0.18.2
geoip2==4.5.0
idna==3.3
importlib-metadata==4.11.3
Markdown==3.3.6
maxminddb==2.2.0
mssql-django==1.1.3
multidict==6.0.2
Pillow==9.1.0
pkg_resources==0.0.0
pycodestyle==2.8.0
pylab-sdk==1.3.2
pyodbc==4.0.32
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
timezones==2.0.9
toml==0.10.2
urllib3==1.26.9
yarl==1.7.2
zipp==3.8.0
